+++ src/main/nativeApplicationMenu.ts
@@ -29,9 +29,9 @@ export function nativeMacApplicationMenuTemplate(
       label: 'Edit',
       submenu: [
         { label: 'Copy', role: 'copy', accelerator: 'CommandOrControl+C' },
+        { label: 'Paste', role: 'paste', accelerator: 'CommandOrControl+V' },
         {
           label: 'Paste Steering',
-          accelerator: 'CommandOrControl+V',
           click: () => callbacks.dispatchRendererAction('paste_steering')
         }
       ]